requirements.txt
faker
pandas
python-dotenv
snowflake-connector-python[pandas]
